requirements.txt
redis==5.2.0
fastapi
uvicorn
openai>=1.50
anthropic>=0.35
instructor>=1.5
pydantic>=2.0
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv
anyio
boto3
pinecone-text
pinecone
pymongo
rich
rapidfuzz